refactored: seveal code groups are extracted to modules. and global variables are aggregated to a special module.
