+++ chord_sim/gval.py
@@ -1,0 +1,43 @@
+# coding:utf-8
+
+import threading
+from chord_node import *
+from chord_util import *
+from typing import Dict, List, Any, Optional, cast
+
+ID_SPACE_BITS = 30 # 160 <- sha1での本来の値
+ID_SPACE_RANGE = 2**ID_SPACE_BITS # 0を含めての数である点に注意
+
+STABILIZE_SUCCESSOR_BATCH_TIMES = 20 #10 #20
+STABILIZE_FTABLE_BATCH_TIMES = 1
+
+# # 160bit符号なし整数の最大値
+# # Chordネットワーク上のID空間の上限
+# ID_MAX = 2**ID_SPACE_BITS - 1
+
+# 30bit符号なし整数の最大値
+# Chordネットワーク上のID空間の上限
+# TODO: 検証時の実行時間短縮のためにハッシュ関数で求めた値の代わりに乱数
+#       を用いているため bit数 を少なくしている
+ID_MAX = ID_SPACE_RANGE - 1
+
+NODE_NUM = 1000
+# PUT_DATA_NUM = 100
+
+# アドレス文字列をキーとしてとり、対応するノードのChordNodeオブジェクトを返すハッシュ
+# IPアドレスが分かれば、対応するノードと通信できることと対応している
+all_node_dict : Dict[str, 'ChordNode'] = {}
+
+# DHT上で保持されている全てのデータが保持されているリスト
+# KeyValueオブジェクトを要素として持つ
+# 全てのノードはputの際はDHTにデータをputするのとは別にこのリストにデータを追加し、
+# getする際はDHTに対してgetを発行するためのデータをこのリストからランダム
+# に選び、そのkeyを用いて探索を行う. また value も保持しておき、取得できた内容と
+# 照らし合わせられるようにする
+all_data_list : List['KeyValue'] = []
+
+# 検証を分かりやすくするために何ノード目として生成されたか
+# のデバッグ用IDを持たせるためのカウンタ
+already_born_node_num = 0
+
+lock_of_all_data = threading.Lock()
